Updating triennial outstanding contributions via celery task (#578)
The celery task will be executed annualy to update the triennial status of contributions. 

This ensures that the `outstanding_contributions` field will only be based on the agreed contributions up to the current year. The `outstanding_contributions` field for each new triennial is initially populated automatically (with values worth a year out of the three) when a new SoA is finalized for that triennial.